vibe coded my way out of feeling unmotivated for this project.
After abandoning it for a month, I'm trying to figure out new ways to
approach this. My pain point the first time around was configuring fastAPI
and dealing with all of the complexities of delving into a new technology
that doesn't do any hand holding while trying to make sense of it all.

Now this time around, I'm trying to generate a good boilerplate to work
with it while keeping things simple and modular.